get rid of _FD_SET usage, instead just case the descriptor as an int.  hopefully this sits well with the windows build.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@10431 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/c-ares/ares_fds.c
@@ -39,15 +39,15 @@ int ares_fds(ares_channel channel, fd_set *read_fds, fd_set *write_fds)
       server = &channel->servers[i];
       if (server->udp_socket != -1)
 	{
-	  FD_SET(server->udp_socket, read_fds);
+	  FD_SET((unsigned int)server->udp_socket, read_fds);
 	  if (server->udp_socket >= nfds)
 	    nfds = server->udp_socket + 1;
 	}
       if (server->tcp_socket != -1)
 	{
-	  FD_SET(server->tcp_socket, read_fds);
+	  FD_SET((unsigned int)server->tcp_socket, read_fds);
 	  if (server->qhead)
-	    FD_SET(server->tcp_socket, write_fds);
+	    FD_SET((unsigned int)server->tcp_socket, write_fds);
 	  if (server->tcp_socket >= nfds)
 	    nfds = server->tcp_socket + 1;
 	}
